Rename files to conform the standard; fix linter errors
