+++ src/__test__/components/Header.test.js
@@ -3,7 +3,7 @@ import { BrowserRouter as Router } from 'react-router-dom';
 import renderer from 'react-test-renderer';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import Header from '../../components/Header/Header';
+import Header from '../../components/Header';
 
 let getByTestId;
