+++ src/containers/CryptoQuote/index.js
@@ -18,7 +18,7 @@ import { getCryptocurrencyQuoteAsync } from '../../redux/cryptoQuotesSlice';
 
 import styles from './CryptoQuote.module.css';
 import mainStyles from '../../index.module.css';
-import ErrorAlert from '../../components/Error/Error';
+import ErrorAlert from '../../components/Error';
 
 const CruyptoQuote = () => {
   // const [cryptoQuote, setCryptoQuote] = useState(null);
