+++ src/containers/Trending/index.js
@@ -19,7 +19,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { getTrendingCryptosAsync } from '../../redux/trendingSlice';
 import styles from './Trending.module.css';
 import mainStyles from '../../index.module.css';
-import ErrorAlert from '../../components/Error/Error';
+import ErrorAlert from '../../components/Error';
 
 const useStyles = makeStyles({
   table: {
@@ -49,7 +49,6 @@ const Trending = () => {
   }, [dispatch]);
 
   const classes = useStyles();
-  console.log(loading);
 
   if (loading) {
     return <WaveLoading />;
